Update Project package name to follow reverse domain format
Closes #14
