+++ .gitignore
@@ -184,4 +184,4 @@ ctre_sim/
 compile_commands.json
 
 # Version file
-src/main/java/frc/robot/BuildConstants.java
+src/main/java/org/jmhsrobotics/frc2025/BuildConstants.java
