+++ src/main/java/org/jmhsrobotics/frc2025/Constants.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot;
+package org.jmhsrobotics.frc2025;
 
 import edu.wpi.first.wpilibj.RobotBase;
