+++ src/main/java/org/jmhsrobotics/frc2025/subsystems/drive/Drive.java
@@ -11,10 +11,10 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.drive;
+package org.jmhsrobotics.frc2025.subsystems.drive;
 
 import static edu.wpi.first.units.Units.*;
-import static frc.robot.subsystems.drive.DriveConstants.*;
+import static org.jmhsrobotics.frc2025.subsystems.drive.DriveConstants.*;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.config.PIDConstants;
@@ -42,13 +42,13 @@ import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
-import frc.robot.Constants;
-import frc.robot.Constants.Mode;
-import frc.robot.subsystems.drive.swerve.ModuleIO;
-import frc.robot.subsystems.drive.swerve.ModuleThrifty;
-import frc.robot.util.LocalADStarAK;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
+import org.jmhsrobotics.frc2025.Constants;
+import org.jmhsrobotics.frc2025.Constants.Mode;
+import org.jmhsrobotics.frc2025.subsystems.drive.swerve.ModuleIO;
+import org.jmhsrobotics.frc2025.subsystems.drive.swerve.ModuleThrifty;
+import org.jmhsrobotics.frc2025.util.LocalADStarAK;
 import org.littletonrobotics.junction.AutoLogOutput;
 import org.littletonrobotics.junction.Logger;
