+++ src/main/java/org/jmhsrobotics/frc2025/subsystems/drive/GyroIONavX.java
@@ -11,9 +11,9 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.drive;
+package org.jmhsrobotics.frc2025.subsystems.drive;
 
-import static frc.robot.subsystems.drive.DriveConstants.*;
+import static org.jmhsrobotics.frc2025.subsystems.drive.DriveConstants.*;
 
 import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.math.geometry.Rotation2d;
