+++ src/main/java/org/jmhsrobotics/frc2025/subsystems/drive/GyroIOPigeon2.java
@@ -11,9 +11,9 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.drive;
+package org.jmhsrobotics.frc2025.subsystems.drive;
 
-import static frc.robot.subsystems.drive.DriveConstants.*;
+import static org.jmhsrobotics.frc2025.subsystems.drive.DriveConstants.*;
 
 import com.ctre.phoenix6.BaseStatusSignal;
 import com.ctre.phoenix6.StatusCode;
