+++ src/main/java/org/jmhsrobotics/frc2025/subsystems/drive/swerve/ModuleIOThrifty.java
@@ -11,10 +11,10 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.drive.swerve;
+package org.jmhsrobotics.frc2025.subsystems.drive.swerve;
 
-import static frc.robot.subsystems.drive.DriveConstants.*;
-import static frc.robot.util.SparkUtil.*;
+import static org.jmhsrobotics.frc2025.subsystems.drive.DriveConstants.*;
+import static org.jmhsrobotics.frc2025.util.SparkUtil.*;
 
 import com.revrobotics.AbsoluteEncoder;
 import com.revrobotics.RelativeEncoder;
@@ -24,6 +24,7 @@ import com.revrobotics.spark.SparkBase.PersistMode;
 import com.revrobotics.spark.SparkBase.ResetMode;
 import com.revrobotics.spark.SparkClosedLoopController;
 import com.revrobotics.spark.SparkClosedLoopController.ArbFFUnits;
+import com.revrobotics.spark.SparkFlex;
 import com.revrobotics.spark.SparkLowLevel.MotorType;
 import com.revrobotics.spark.SparkMax;
 import com.revrobotics.spark.config.ClosedLoopConfig.FeedbackSensor;
@@ -33,16 +34,16 @@ import com.revrobotics.spark.config.SparkMaxConfig;
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.filter.Debouncer;
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.subsystems.drive.DriveConstants.revConstants;
-import frc.robot.subsystems.drive.SparkOdometryThread;
 import java.util.Queue;
 import java.util.function.DoubleSupplier;
+import org.jmhsrobotics.frc2025.subsystems.drive.DriveConstants.thriftyConstants;
+import org.jmhsrobotics.frc2025.subsystems.drive.SparkOdometryThread;
 
 /**
  * Module IO implementation for Spark Flex drive motor controller, Spark Max turn motor controller,
  * and duty cycle absolute encoder.
  */
-public class ModuleIORev implements ModuleIO {
+public class ModuleIOThrifty implements ModuleIO {
   private final Rotation2d zeroRotation;
 
   // Hardware objects
@@ -64,7 +65,7 @@ public class ModuleIORev implements ModuleIO {
   private final Debouncer driveConnectedDebounce = new Debouncer(0.5);
   private final Debouncer turnConnectedDebounce = new Debouncer(0.5);
 
-  public ModuleIORev(int module) {
+  public ModuleIOThrifty(int module) {
     zeroRotation =
         switch (module) {
           case 0 -> frontLeftZeroRotation;
@@ -74,22 +75,22 @@ public class ModuleIORev implements ModuleIO {
           default -> new Rotation2d();
         };
     driveSpark =
-        new SparkMax(
+        new SparkFlex(
             switch (module) {
-              case 0 -> revConstants.frontLeftDriveCanId;
-              case 1 -> revConstants.frontRightDriveCanId;
-              case 2 -> revConstants.backLeftDriveCanId;
-              case 3 -> revConstants.backRightDriveCanId;
+              case 0 -> thriftyConstants.frontLeftDriveCanId;
+              case 1 -> thriftyConstants.frontRightDriveCanId;
+              case 2 -> thriftyConstants.backLeftDriveCanId;
+              case 3 -> thriftyConstants.backRightDriveCanId;
               default -> 0;
             },
             MotorType.kBrushless);
     turnSpark =
         new SparkMax(
             switch (module) {
-              case 0 -> revConstants.frontLeftTurnCanId;
-              case 1 -> revConstants.frontRightTurnCanId;
-              case 2 -> revConstants.backLeftTurnCanId;
-              case 3 -> revConstants.backRightTurnCanId;
+              case 0 -> thriftyConstants.frontLeftTurnCanId;
+              case 1 -> thriftyConstants.frontRightTurnCanId;
+              case 2 -> thriftyConstants.backLeftTurnCanId;
+              case 3 -> thriftyConstants.backRightTurnCanId;
               default -> 0;
             },
             MotorType.kBrushless);
@@ -102,20 +103,20 @@ public class ModuleIORev implements ModuleIO {
     var driveConfig = new SparkFlexConfig();
     driveConfig
         .idleMode(IdleMode.kBrake)
-        .smartCurrentLimit(revConstants.driveMotorCurrentLimit)
+        .smartCurrentLimit(thriftyConstants.driveMotorCurrentLimit)
         .voltageCompensation(12.0);
     driveConfig
         .encoder
-        .positionConversionFactor(revConstants.driveEncoderPositionFactor)
-        .velocityConversionFactor(revConstants.driveEncoderVelocityFactor)
+        .positionConversionFactor(thriftyConstants.driveEncoderPositionFactor)
+        .velocityConversionFactor(thriftyConstants.driveEncoderVelocityFactor)
         .uvwMeasurementPeriod(10)
         .uvwAverageDepth(2);
     driveConfig
         .closedLoop
         .feedbackSensor(FeedbackSensor.kPrimaryEncoder)
         .pidf(
-            revConstants.driveKp, 0.0,
-            revConstants.driveKd, 0.0);
+            thriftyConstants.driveKp, 0.0,
+            thriftyConstants.driveKd, 0.0);
     driveConfig
         .signals
         .primaryEncoderPositionAlwaysOn(true)
@@ -136,22 +137,23 @@ public class ModuleIORev implements ModuleIO {
     // Configure turn motor
     var turnConfig = new SparkMaxConfig();
     turnConfig
-        .inverted(revConstants.turnInverted)
+        .inverted(thriftyConstants.turnInverted)
         .idleMode(IdleMode.kBrake)
-        .smartCurrentLimit(revConstants.turnMotorCurrentLimit)
+        .smartCurrentLimit(thriftyConstants.turnMotorCurrentLimit)
         .voltageCompensation(12.0);
     turnConfig
         .absoluteEncoder
-        .inverted(revConstants.turnEncoderInverted)
-        .positionConversionFactor(revConstants.turnEncoderPositionFactor)
-        .velocityConversionFactor(revConstants.turnEncoderVelocityFactor)
+        .inverted(thriftyConstants.turnEncoderInverted)
+        .positionConversionFactor(thriftyConstants.turnEncoderPositionFactor)
+        .velocityConversionFactor(thriftyConstants.turnEncoderVelocityFactor)
         .averageDepth(2);
     turnConfig
         .closedLoop
         .feedbackSensor(FeedbackSensor.kAbsoluteEncoder)
         .positionWrappingEnabled(true)
-        .positionWrappingInputRange(revConstants.turnPIDMinInput, revConstants.turnPIDMaxInput)
-        .pidf(revConstants.turnKp, 0.0, revConstants.turnKd, 0.0);
+        .positionWrappingInputRange(
+            thriftyConstants.turnPIDMinInput, thriftyConstants.turnPIDMaxInput)
+        .pidf(thriftyConstants.turnKp, 0.0, thriftyConstants.turnKd, 0.0);
     turnConfig
         .signals
         .absoluteEncoderPositionAlwaysOn(true)
@@ -230,8 +232,8 @@ public class ModuleIORev implements ModuleIO {
   @Override
   public void setDriveVelocity(double velocityRadPerSec) {
     double ffVolts =
-        revConstants.driveKs * Math.signum(velocityRadPerSec)
-            + revConstants.driveKv * velocityRadPerSec;
+        thriftyConstants.driveKs * Math.signum(velocityRadPerSec)
+            + thriftyConstants.driveKv * velocityRadPerSec;
     driveController.setReference(
         velocityRadPerSec, ControlType.kVelocity, 0, ffVolts, ArbFFUnits.kVoltage);
   }
@@ -241,8 +243,8 @@ public class ModuleIORev implements ModuleIO {
     double setpoint =
         MathUtil.inputModulus(
             rotation.plus(zeroRotation).getRadians(),
-            revConstants.turnPIDMinInput,
-            revConstants.turnPIDMaxInput);
+            thriftyConstants.turnPIDMinInput,
+            thriftyConstants.turnPIDMaxInput);
     turnController.setReference(setpoint, ControlType.kPosition);
   }
 }
