+++ src/main/java/org/jmhsrobotics/frc2025/subsystems/drive/swerve/ModuleThrifty.java
@@ -11,17 +11,18 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.subsystems.drive.swerve;
+package org.jmhsrobotics.frc2025.subsystems.drive.swerve;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.wpilibj.Alert;
 import edu.wpi.first.wpilibj.Alert.AlertType;
-import frc.robot.subsystems.drive.DriveConstants.revConstants;
+import org.jmhsrobotics.frc2025.subsystems.drive.DriveConstants.thriftyConstants;
+// import frc.robot.subsystems.drive.ModuleIOInputsAutoLogged;
 import org.littletonrobotics.junction.Logger;
 
-public class ModuleRev {
+public class ModuleThrifty {
   private final ModuleIO io;
   private final ModuleIOInputsAutoLogged inputs = new ModuleIOInputsAutoLogged();
   private final int index;
@@ -30,7 +31,7 @@ public class ModuleRev {
   private final Alert turnDisconnectedAlert;
   private SwerveModulePosition[] odometryPositions = new SwerveModulePosition[] {};
 
-  public ModuleRev(ModuleIO io, int index) {
+  public ModuleThrifty(ModuleIO io, int index) {
     this.io = io;
     this.index = index;
     driveDisconnectedAlert =
@@ -50,7 +51,8 @@ public class ModuleRev {
     int sampleCount = inputs.odometryTimestamps.length; // All signals are sampled together
     odometryPositions = new SwerveModulePosition[sampleCount];
     for (int i = 0; i < sampleCount; i++) {
-      double positionMeters = inputs.odometryDrivePositionsRad[i] * revConstants.wheelRadiusMeters;
+      double positionMeters =
+          inputs.odometryDrivePositionsRad[i] * thriftyConstants.wheelRadiusMeters;
       Rotation2d angle = inputs.odometryTurnPositions[i];
       odometryPositions[i] = new SwerveModulePosition(positionMeters, angle);
     }
@@ -67,7 +69,7 @@ public class ModuleRev {
     state.cosineScale(inputs.turnPosition);
 
     // Apply setpoints
-    io.setDriveVelocity(state.speedMetersPerSecond / revConstants.wheelRadiusMeters);
+    io.setDriveVelocity(state.speedMetersPerSecond / thriftyConstants.wheelRadiusMeters);
     io.setTurnPosition(state.angle);
   }
 
@@ -90,12 +92,12 @@ public class ModuleRev {
 
   /** Returns the current drive position of the module in meters. */
   public double getPositionMeters() {
-    return inputs.drivePositionRad * revConstants.wheelRadiusMeters;
+    return inputs.drivePositionRad * thriftyConstants.wheelRadiusMeters;
   }
 
   /** Returns the current drive velocity of the module in meters per second. */
   public double getVelocityMetersPerSec() {
-    return inputs.driveVelocityRadPerSec * revConstants.wheelRadiusMeters;
+    return inputs.driveVelocityRadPerSec * thriftyConstants.wheelRadiusMeters;
   }
 
   /** Returns the module position (turn angle and drive position). */
