+++ src/main/java/org/jmhsrobotics/frc2025/util/LocalADStarAK.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package org.jmhsrobotics.frc2025.util;
 
 import com.pathplanner.lib.path.GoalEndState;
 import com.pathplanner.lib.path.PathConstraints;
