+++ src/main/java/org/jmhsrobotics/frc2025/util/SparkUtil.java
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
 
-package frc.robot.util;
+package org.jmhsrobotics.frc2025.util;
 
 import com.revrobotics.REVLibError;
 import com.revrobotics.spark.SparkBase;
